chore: Rename Testcontainers for C to Testcontainers Native
This renames every instance of "Testcontainers for C", "Testcontainers C", and "testcontainers-c".
