+++ CMakeLists.txt
@@ -1,11 +1,11 @@
 cmake_minimum_required (VERSION 3.16.3)
-project (TESTCONTAINERS-C)
+project (TESTCONTAINERS-NATIVE)
 
 include(CTest)
 
-include_directories(${CMAKE_CURRENT_BINARY_DIR}/testcontainers-c)
+include_directories(${CMAKE_CURRENT_BINARY_DIR}/testcontainers-native)
 
-add_subdirectory(testcontainers-c)
+add_subdirectory(testcontainers-native)
 add_subdirectory(modules)
 if(NOT DEFINED SKIP_DEMOS)
     add_subdirectory(demo)
