+++ ROADMAP.md
@@ -3,23 +3,23 @@
 [![Slack: testcontainers-native on slack.testcontainers.org](https://img.shields.io/badge/Slack-%23testcontainers%E2%80%94native-brightgreen?style=flat&logo=slack)](http://slack.testcontainers.org/)
 
 A full public roadmap is coming soon.
-See the [GitHub Issues](https://github.com/testcontainers/testcontainers-c/issues) and,
+See the [GitHub Issues](https://github.com/testcontainers/testcontainers-native/issues) and,
 in particular,
 the pinned issues to learn about the key initiatives on the table.
 
 ## Key Objectives
 
-What we want to do before 1.0: 
+What we want to do before 1.0:
 
-- [ ] Ensure sufficient feature parity with Testcontainers for Go ([#6](https://github.com/testcontainers/testcontainers-c/issues/6))
-- [ ] Add a C++ binding library and class-based headers ([#12](https://github.com/testcontainers/testcontainers-c/issues/12))
+- [ ] Ensure sufficient feature parity with Testcontainers for Go ([#6](https://github.com/testcontainers/testcontainers-native/issues/6))
+- [ ] Add a C++ binding library and class-based headers ([#12](https://github.com/testcontainers/testcontainers-native/issues/12))
 - [ ] Prototype and document support for other languages that
   leverage native libraries, in particular Swift
 - [x] Make Testcontainers Native an official project
   listed on the [Testcontainers site](https://testcontainers.com/)
 - [ ] Publish the project to common C/C++ package managers
-  ([vcpkg](https://github.com/testcontainers/testcontainers-c/issues/2),
-  [Conan](https://github.com/testcontainers/testcontainers-c/issues/3))
+  ([vcpkg](https://github.com/testcontainers/testcontainers-native/issues/2),
+  [Conan](https://github.com/testcontainers/testcontainers-native/issues/3))
 
 ## Discuss the Roadmap
