+++ demo/README.md
@@ -3,5 +3,5 @@
 ## Usage in C/C++
 
 - [Using the generic Testcontainer C API](./generic-container/README.md)
-- [Using Testcontainers C in Google Test (C++)](./google-test/README.md)
+- [Using Testcontainers Native in Google Test (C++)](./google-test/README.md)
 - [Using the WireMock module](./wiremock/README.md)
