+++ demo/generic-container/README.md
@@ -1,6 +1,6 @@
-# Demo - Generic Container on Testcontainers C
+# Demo - Generic Container on Testcontainers Native
 
-Demonstrates usage of the generic container API provided by the Testcontainers C library.
+Demonstrates usage of the generic container API provided by the Testcontainers Native library.
 In this demo, we do not use any module SDKs or test frameworks, just a simple `main()` function.
 
 ## Running demo
