+++ demo/generic-container/generic_container_demo.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
-#include "testcontainers-c.h"
+#include "testcontainers-native.h"
 
 #define DEFAULT_IMAGE "wiremock/wiremock:3.0.1-1"
 
 int main() {
-    printf("Using WireMock with the Testcontainers C binding:\n");
+    printf("Using WireMock with the Testcontainers Native binding:\n");
 
     printf("Creating new container: %s\n", DEFAULT_IMAGE);
     int requestId = tc_new_container_request(DEFAULT_IMAGE);
@@ -31,4 +31,3 @@ int main() {
     printf("Server Response: HTTP-%d\n%s\n\n", response.r0, response.r1);
     return 0;
 }
-
