+++ docs/cpp/README.md
@@ -2,7 +2,7 @@
 
 At the moment, there is no dedicated C++ binding library/header,
 but it is on [our roadmap](../../ROADMAP.md).
-Tou can use the `testcontainers-c` library directly
+Tou can use the `testcontainers-native` library directly
 in all C++ testing frameworks.
 
 ## Google Test
@@ -15,7 +15,7 @@ with proper setup and tear-down methods.
 #include <iostream>
 #include <string>
 #include <gtest/gtest.h>
-#include "testcontainers-c.h"
+#include "testcontainers-native.h"
 
 class WireMockTestContainer : public ::testing::Test {
 
@@ -29,10 +29,10 @@ protected:
         tc_with_exposed_tcp_port(requestId, 8080);
         tc_with_wait_for_http(requestId, 8080, WIREMOCK_ADMIN_MAPPING_ENDPOINT);
         tc_with_file(requestId, "test_data/hello.json", "/home/wiremock/mappings/hello.json");
-        
+
         struct tc_run_container_return ret = tc_run_container(requestId);
         containerId = ret.r0;
-        
+
         EXPECT_TRUE(ret.r1) << "Failed to run the container: " << ret.r2;
     };
 
@@ -51,14 +51,14 @@ Then, you can define new tests by referring to the container via `containerId`.
 TEST_F(WireMockTestContainer, HelloWorld) {
     std::cout << "Sending HTTP request to the container\n";
     struct tc_send_http_get_return response = tc_send_http_get(containerId, 8080, "/hello");
-    
+
     ASSERT_NE(response.r0, -1) << "Failed to send HTTP request: " << response.r2;
     ASSERT_EQ(response.r0, 200) << "Received wrong response code: " << response.r1 << response.r2;
- 
+
     std::cout << "Server Response: HTTP-" << response.r0 << '\n' << response.r1 << '\n';
 }
 ```
 
 ## Demos
 
-- [Using Testcontainers C in Google Test (C++)](../../demo/google-test/)
+- [Using Testcontainers Native in Google Test (C++)](../../demo/google-test/)
