+++ modules/wiremock/CMakeLists.txt
@@ -1,21 +1,21 @@
 
 cmake_minimum_required(VERSION 3.9)
-project(testcontainers-c-wiremock VERSION 0.0.1
-    DESCRIPTION "WireMock module for Testcontainers C")
+project(testcontainers-native-wiremock VERSION 0.0.1
+    DESCRIPTION "WireMock module for Testcontainers Native")
 
 include(GNUInstallDirs)
 
 add_library(${PROJECT_NAME} SHARED
-    testcontainers-c-wiremock.h
+    testcontainers-native-wiremock.h
     impl.c
 )
-add_dependencies(${PROJECT_NAME} testcontainers-c)
+add_dependencies(${PROJECT_NAME} testcontainers-native)
 
-include_directories(${testcontainers-c_SOURCE_DIR})
+include_directories(${testcontainers-native_SOURCE_DIR})
 
-set_target_properties(${PROJECT_NAME} PROPERTIES 
-    VERSION ${PROJECT_VERSION}    
-    PUBLIC_HEADER testcontainers-c-wiremock.h)
+set_target_properties(${PROJECT_NAME} PROPERTIES
+    VERSION ${PROJECT_VERSION}
+    PUBLIC_HEADER testcontainers-native-wiremock.h)
 
 configure_file(cmake.pc.in ${PROJECT_NAME}.pc @ONLY)
 install(TARGETS ${PROJECT_NAME}
