+++ modules/wiremock/README.md
@@ -1,4 +1,4 @@
-# WireMock Module for Testcontainers C
+# WireMock Module for Testcontainers Native
 
 This module allows provisioning the WireMock server as a standalone container within your tests,
 based on [WireMock Docker](https://github.com/wiremock/wiremock-docker).
